[Fix] Router
We accept methods with method request implicit on Controllers.
For example, route /example can be configured to hold controllers: exampleGet, examplePost, examplePut and exampleDelete for the 4 method request.
And it accepts the alias 'example' for Get only.

In other way, a route /only-post configured to method examplePost will not work for any other request if there is not 'exampleGet', 'examplePut' and 'exampleDelete' methods.